first attempt at round robin
i am creating too many go routines to simulate the requests. the amount
of go routines competing for resources eventually causes them to starve
each other of resources. Next step is to try simulating requests by just
keeping track how many  requests are open at a time. and constantly
decrement the number to simulate the the completion of a request.